backend2/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
httpx==0.25.2
aiofiles==23.2.1
redis==5.0.1
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.2.0